Add files via upload
This project is about BangorTelco, a leading provider of mobile phone services, which is tackling the challenge of high customer churn. The project aims to develop a predictive model to identify customers likely to leave at the end of their contracts. This targeted approach will optimize incentive offerings and reduce costs.